Fixed tricky issue to do with quoting policies.
